Rechner fertig; double hinzugefügt, für grössere Operationen

+++ Minirechner/Minirechner/Program.cs
@@ -9,15 +9,15 @@ namespace Minirechner
     {
         static void Main(string[] args)
         {
-            int Zahleins = 0;
-            int Zahlzwei = 0;
+            double Zahleins = 0;
+            double Zahlzwei = 0;
             int Operation = 0;
-            int Ergebnis = 0;
+            double Ergebnis = 0;
 
             Console.WriteLine("Gib die erste Zahl ein:");
-            Zahleins = Convert.ToInt16(Console.ReadLine());
+            Zahleins = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Gib die zweite Zahl ein");
-            Zahlzwei = Convert.ToInt16(Console.ReadLine());
+            Zahlzwei = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Wähle: 1 = Addieren, 2 = Subtrahieren, 3 = Multiplizieren, 4 = Dividieren");
             Operation = Convert.ToInt16(Console.ReadLine());
 
@@ -37,24 +37,24 @@ namespace Minirechner
             Console.ReadKey();
         }
 
-        static int Addieren(int zahl1, int zahl2)
+        static double Addieren(double zahl1, double zahl2)
         {
-            int Erg = zahl1 + zahl2;
+            double Erg = zahl1 + zahl2;
             return Erg;
         }
-        static int Subtrahieren(int zahl1, int zahl2)
+        static double Subtrahieren(double zahl1, double zahl2)
         {
-            int Erg = zahl1 - zahl2;
+            double Erg = zahl1 - zahl2;
             return Erg;
         }
-        static int Multiplizieren(int zahl1, int zahl2)
+        static double Multiplizieren(double zahl1, double zahl2)
         {
-            int Erg = zahl1 * zahl2;
+            double Erg = zahl1 * zahl2;
             return Erg;
         }
-        static int Dividieren(int zahl1, int zahl2)
+        static double Dividieren(double zahl1, double zahl2)
         {
-            int Erg = zahl1 / zahl2;
+            double Erg = zahl1 / zahl2;
             return Erg;
         }
